Cambio de double a int

+++ src/AbstractNode.java
@@ -8,19 +8,19 @@ public abstract class AbstractNode implements INode{
     }
     public AbstractNode() {
     }
-    public Double getMinX() {
+    public Integer getMinX() {
         return rectangle.getMinX();
     }
-    public Double getMaxX() {
+    public Integer getMaxX() {
         return rectangle.getMaxX();
     }
-    public Double getMinY() {
+    public Integer getMinY() {
         return rectangle.getMinY();
     }
-    public Double getMaxY() {
+    public Integer getMaxY() {
         return rectangle.getMaxY();
     }
-    public Double centerX() {
+    public Integer centerX() {
         return rectangle.centerX();
     }
 
@@ -28,7 +28,7 @@ public abstract class AbstractNode implements INode{
         return rectangle.intersect(r);
     }
 
-    public Double centerY() {
+    public Integer centerY() {
         return rectangle.centerY();
     }
